file renaming and added license headers on new files
